[feat] BusinessExceptionHandler 사용 배제, 댓글 삭제 시 검증 기능 추가
1. BusinessExceptionHandler 대신 사용자의 인식이 필요한 상황에 적절한 예외로 변경

2. 댓글 삭제 기능에서 댓글 삭제 전 댓글이 존재하는지 여부와 자식 댓글이 존재하면 삭제 할 수 없도록 기능 추가

+++ src/main/java/com/opus/comment/service/CommentService.java
@@ -1,17 +1,20 @@
 package com.opus.comment.service;
 
 import com.opus.comment.domain.CommentResponseDTO;
+import com.opus.comment.domain.CommentUpdateDTO;
 import com.opus.utils.SecurityUtil;
 import com.opus.comment.domain.CommentDTO;
 import com.opus.comment.domain.CommentVO;
 import com.opus.comment.mapper.CommentMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.NoSuchElementException;
 
 @Service
 @RequiredArgsConstructor
@@ -35,26 +38,23 @@ public class CommentService {
         commentMapper.insertComment(comment);
     }
 
-    // 게시글에 달린 댓글 리스트
     @Transactional(readOnly = true)
     public List<CommentResponseDTO> getCommentListByPinId(int pinId) {
-         List<CommentVO> comments = commentMapper.selectCommentsByPinId(pinId);
-         return comments.stream()
-                 .map(CommentResponseDTO::of)
-                 .collect(Collectors.toList());
+        List<CommentVO> comments = commentMapper.selectCommentsByPinId(pinId);
+        return comments.stream()
+                .map(CommentResponseDTO::of)
+                .toList();
     }
 
-    // 내가 쓴 댓글 리스트
     @Transactional(readOnly = true)
     public List<CommentResponseDTO> getMyCommentList() {
         List<CommentVO> comments = commentMapper.selectCommentsByMemberId(SecurityUtil.getCurrentUserId());
         return comments.stream()
                 .map(CommentResponseDTO::of)
-                .collect(Collectors.toList());
+                .toList();
     }
 
-    // 댓글 수정
-    public void editComment(CommentDTO commentDTO) {
+    public void editComment(CommentUpdateDTO commentDTO) {
         CommentVO comment = CommentVO.builder()
                 .commentId(commentDTO.getCommentId())
                 .pinId(commentDTO.getPinId())
@@ -68,13 +68,23 @@ public class CommentService {
         commentMapper.updateComment(comment);
     }
 
-    // 댓글 삭제
+    @Transactional
     public void removeComment(int commentId) {
+
+        commentMapper.selectCommentByCommentId(commentId)
+                .orElseThrow(() -> new NoSuchElementException("해당 댓글이 존재하지 않습니다."));
+
+
+        if (commentMapper.countChildCommentsByCommentId(commentId) > 0) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "자식 댓글이 존재하는 댓글은 삭제할 수 없습니다.");
+        }
+
+        // TODO 삭제 권한 검증 추가
+
         CommentVO comment = CommentVO.builder()
                 .commentId(commentId)
                 .memberId(SecurityUtil.getCurrentUserId())
                 .build();
         commentMapper.deleteComment(comment);
     }
-
 }
